+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 3.1.0 2022-07-07
 ==================
-  * added new API `FormattingOptions.keepLines` : It leaves the lines as is
+  * added new API `FormattingOptions.keepLines` : It leaves the initial line positions in the formatting
 
 3.0.0 2020-11-13
 ==================